requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
numpy==1.20.0
onnx==1.9.0
onnxruntime==1.7.0
Pillow==8.2.0
uwsgi==2.0.19.1